Frontend update
Added angular material for MatDialog.
Added new grid-* to Tailwindcss.
Created confirmation dialog and confirmation dialog service.
Implemented filtered search functionality, added new DateFilter model to models.
Changed UI design for profile page and search bar component.
Changed how popup type is passed to image service.
Changed some route names in app.routes.ts.

+++ frontend/src/app/admin-page/admin-data-row-style/admin-data-row-style.component.ts
@@ -4,6 +4,7 @@ import { UserService } from '../../services/user.service';
 import { PopupService } from '../../popups/popup.service';
 import { EditUserComponentService } from '../../edit-user/edit-user-component.service';
 import { RouterLink } from '@angular/router';
+import ConfirmationDialogService from '../../confirmation-dialog/confirmation-dialog.service';
 
 @Component({
   selector: 'app-admin-data-row-style',
@@ -21,17 +22,26 @@ export class AdminDataRowStyleComponent {
   userService : UserService = inject(UserService);
   popupService : PopupService = inject(PopupService);
   editUserComponentService : EditUserComponentService = inject(EditUserComponentService);
+  confirmationDialogService : ConfirmationDialogService = inject(ConfirmationDialogService);
   destroyRef : DestroyRef = inject(DestroyRef);
 
   DeleteUser(){
-    if(this.adminService.UserIsAdmin() && this.userService.GetUsername() == this.data.username){
-      this.popupService.ShowPopup("Admin felhasználó nem törölheti önmagát!", "warning");
-      return;
-    }
-    const subscription = this.adminService.DeleteUser(this.data.ID!).subscribe();
+    const confirmationDialogServiceSubscription = this.confirmationDialogService.OpenDialog("Biztosan törölni szeretnéd ezt a felhasználót?").subscribe(result => {
+      if(result == "ok"){
+        if(this.adminService.UserIsAdmin() && this.userService.GetUsername() == this.data.username){
+          this.popupService.ShowPopup("Admin felhasználó nem törölheti önmagát!", "warning");
+          return;
+        }
+        const subscription = this.adminService.DeleteUser(this.data.ID!).subscribe();
+    
+        this.destroyRef.onDestroy(() => {
+          subscription.unsubscribe();
+        });
+      }
+    });
 
     this.destroyRef.onDestroy(() => {
-      subscription.unsubscribe();
+      confirmationDialogServiceSubscription.unsubscribe();
     });
   }
